+++ client/src/lib/navigation/SessionStack.tsx
@@ -1,0 +1,32 @@
+import React from 'react';
+import {
+  createNativeStackNavigator,
+  NativeStackNavigationOptions,
+} from '@react-navigation/native-stack';
+import {SessionStackProps} from './constants/routes';
+import Session from '../../routes/Session/Session';
+import ChangingRoom from '../../routes/Session/ChangingRoom';
+import IntroPortal from '../../routes/Session/IntroPortal';
+import OutroPortal from '../../routes/Session/OutroPortal';
+import DailyProvider from '../daily/DailyProvider';
+
+const SessionStack = createNativeStackNavigator<SessionStackProps>();
+
+const screenOptions: NativeStackNavigationOptions = {
+  animation: 'fade',
+  animationDuration: 2000,
+  gestureEnabled: false,
+};
+
+const SessionStackWrapper = () => (
+  <DailyProvider>
+    <SessionStack.Navigator screenOptions={screenOptions}>
+      <SessionStack.Screen name={'ChangingRoom'} component={ChangingRoom} />
+      <SessionStack.Screen name={'IntroPortal'} component={IntroPortal} />
+      <SessionStack.Screen name={'Session'} component={Session} />
+      <SessionStack.Screen name={'OutroPortal'} component={OutroPortal} />
+    </SessionStack.Navigator>
+  </DailyProvider>
+);
+
+export default SessionStackWrapper;
